requirements.txt
tensorflow==2.12.0
pandas
dvc
mlflow==2.2.2
notebook
numpy
matplotlib
seaborn
python-box==6.0.2 # for easy dictionary access
pyYAML 
tqdm
ensure==1.0.2  # ensure annotations to check for correct data types in functions
joblib
types-PyYAML
scipy
Flask
Flask-Cors
kagglehub